Move SubprocessProcessHandler near its single use.
The setup-py goal discovered that process_handler.py was no longer used
by production code. Move it where its used and kill the unused
abstraction.

[ci skip-rust-tests]
