+++ src/python/pants/testutil/BUILD
@@ -29,6 +29,7 @@ python_library(
   sources = ['pants_run_integration_test.py'],
   dependencies = [
     '//:build_root',
+    ':process_handler',
     'src/python/pants:entry_point'
   ],
 )
@@ -46,6 +47,7 @@ target(
   ],
 )
 
+python_tests(name='tests')
 
 python_library(
   name = 'test_base',
@@ -56,7 +58,6 @@ python_library(
   name = 'external_tool_test_base',
   sources = ['external_tool_test_base.py'],
 )
-
 
 python_library(
   name = 'goal_rule_test_base',
@@ -81,6 +82,11 @@ python_library(
 python_library(
   name='pexrc_util',
   sources=['pexrc_util.py'],
+)
+
+python_library(
+  name='process_handler',
+  sources=['process_handler.py'],
 )
 
 python_library(
